requirements.txt
Django>=4.2
psycopg2-binary
python-dotenv
django-ninja==1.3.0
PyJWT==2.8.0
bcrypt==4.1.2